garv-amadeus-text2sql/app/requirements_mcp.txt
# MCP server env (fastmcp needs websockets>=15)
mcp>=1.25.0
fastmcp>=2.14.2
uvicorn>=0.35
starlette>=0.50.0
python-dotenv>=1.1.0
PyYAML>=6.0.1
duckdb>=1.1.0
pandas>=2.2.0
requests>=2.31.0